add TestFunc tool

+++ DriverMonitor/dm_global.cpp
@@ -87,24 +87,24 @@ int removeContents(const std::string &path){
     assert(NULL!=dir);
     struct dirent *ptr=readdir(dir);
     while(NULL!=ptr){
-        //std::cout<<ptr->d_name<<std::endl;
-        if(8==ptr->d_type){
-            if(-1==remove((path+"/"+ptr->d_name).c_str())){
-                return -1;
+        if(0!=strcmp(ptr->d_name,".")&&0!=strcmp(ptr->d_name,"..")){
+            if(4==ptr->d_type){
+                if(-1==removeContents((path+"/"+ptr->d_name).c_str())|| -1==rmdir((path+"/"+ptr->d_name).c_str())){
+                      return -1;
+                }
+            }else{
+                if(-1==remove((path+"/"+ptr->d_name).c_str())){
+                    return -1;
+                }
             }
-        }else if(4==ptr->d_type){
-          if(-1==removeContents((path+"/"+ptr->d_name).c_str())|| -1==rmdir((path+"/"+ptr->d_name).c_str())){
-                return -1;
-          }
         }
         ptr=readdir(dir);
     }
     return 0;
 }
 
-
 //check if folder exists,if not ,make the folder
-void checkFolder(const std::string &folderName){
+void checkFolder(const std::string &folderName,bool bRetain){
     //determine if the folder exists
     if(-1==access(folderName.c_str(),F_OK)){
         //folder doesn't exists
@@ -112,6 +112,10 @@ void checkFolder(const std::string &folderName){
             //fail to mkdir folder
                 std::cout<<"fail to mkdir "<<folderName.c_str()<<std::endl;
                 exit(-1);
+        }
+    }else{
+        if(!bRetain){
+            removeContents(folderName);
         }
     }
 }
